backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
click==8.1.8
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
typing-inspection==0.4.0
typing_extensions==4.12.0
uvicorn==0.27.1
databases==0.8.0
asyncpg==0.29.0
email-validator==2.1.0.post1
SQLAlchemy==1.4.49
databases==0.8.0
